preflight|refactor: Extract `resolveConfigExports` and simplify `-o` flag parsing
Extract the duplicated config-resolution logic (checklists extraction, default-export fallback, `fixLocation` merge) from `config.ts` and `loadRemoteConfig.ts` into a shared `resolveConfigExports` helper. Handle the `-o` short alias directly in the arg loop instead of normalizing it to `--output` before passing to `extractFlagValue`.
